pyproject.toml
[project]
name = "ResortBookingSystem"
version = "0.1.0"
description = "A full-stack resort booking and management website using Python"
dependencies = [
    "Flask==3.1.1",
    "Flask-SQLAlchemy==3.1.1",
    "Flask-Login==0.6.3",
    "Flask-Mail==0.10.0",
    "Flask-WTF==1.2.2",
    "email-validator==2.2.0",
    "Werkzeug==3.1.0",
    "python-dotenv==1.0.1",
    "stripe==9.0.0",
    "gunicorn==23.0.0",
    "SQLAlchemy==2.0.29",
    "bcrypt==4.1.2",
    "PyMySQL==1.1.0",
]
requires-python = ">= 3.11"
license = "MIT"
